Fix #5: declare `gnumake` executable.
This commit fixes #5 by setting `executable = True` in the [`genrule`] call.

This target has been renamed to `gnumake`, and only the executable is exposed,
since it does not rely on any extra data files.

The `GNUMakeToolchainInfo` has been modified to take a [`RunInfo`] in order
to restrict the binary type.

[`genrule`]: https://buck2.build/docs/api/rules/#genrule
[`RunInfo`]: https://buck2.build/docs/api/build/globals/#runinfo